# swagger/v1/swagger.yaml
---
openapi: 3.0.1
info:
  title: API V1
  version: v1
paths:
  "/api/v1/url_shorteners":
    post:
      summary: Creates a shortened URL
      tags:
      - UrlShorteners
      security:
      - BearerAuth: []
      parameters: []
      responses:
        '201':
          description: shortened URL created
        '401':
          description: unauthorized
        '422':
          description: invalid URL
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                original:
                  type: string
                  example: https://localhost:8020
              required:
              - original
    get:
      summary: Lists all shortened URLs
      tags:
      - UrlShorteners
      responses:
        '200':
          description: successful retrieval of shortened URLs
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      example: 1
                    original:
                      type: string
                      example: https://example.com
                    short:
                      type: string
                      example: http://localhost:3000/M2rckK
                    created_at:
                      type: string
                      example: '2024-11-27T01:00:00Z'
                    updated_at:
                      type: string
                      example: '2024-11-27T01:00:00Z'
        '404':
          description: no URL shorteners found
  "/api/v1/url_shorteners/{id}":
    put:
      summary: Updates a shortened URL
      tags:
      - UrlShorteners
      parameters:
      - name: id
        in: path
        description: ID of the URL Shortener
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: shortened URL updated
        '404':
          description: URL Shortener not found
        '422':
          description: validation errors
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                original_url:
                  type: string
                  example: https://updated-example.com
              required:
              - original_url
    delete:
      summary: Deletes a shortened URL
      tags:
      - UrlShorteners
      parameters:
      - name: id
        in: path
        description: ID of the URL Shortener
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: URL Shortener deleted
        '404':
          description: URL Shortener not found
        '422':
          description: failed to delete URL Shortener
    get:
      summary: Fetches details of a specific shortened URL
      tags:
      - UrlShorteners
      parameters:
      - name: id
        in: path
        description: ID of the URL Shortener
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: URL Shortener details fetched successfully
        '404':
          description: URL Shortener not found
servers:
- url: http://{defaultHost}
  variables:
    defaultHost:
      default: localhost:3000
components:
  securitySchemes:
    BearerAuth:
      type: apiKey
      in: header
      name: Authorization
      description: Authorization token (e.g., Bearer <your_token>)
security:
- BearerAuth: []